Original change: https://android-review.googlesource.com/c/platform/tools/tradefederation/+/2040907

+++ src/com/android/tradefed/device/NativeDevice.java
@@ -1243,42 +1243,7 @@ public class NativeDevice implements IManagedTestDevice, IConfigurationReceiver 
                     return handler.pullFile(remoteFilePath, localFile);
                 }
             }
-
-            DeviceAction pullAction =
-                    new DeviceAction() {
-                        @Override
-                        public boolean run()
-                                throws TimeoutException, IOException, AdbCommandRejectedException,
-                                        SyncException {
-                            SyncService syncService = null;
-                            boolean status = false;
-                            try {
-                                syncService = getIDevice().getSyncService();
-                                syncService.pullFile(
-                                        interpolatePathVariables(remoteFilePath),
-                                        localFile.getAbsolutePath(),
-                                        SyncService.getNullProgressMonitor());
-                                status = true;
-                            } catch (SyncException e) {
-                                CLog.w(
-                                        "Failed to pull %s from %s to %s. Message %s",
-                                        remoteFilePath,
-                                        getSerialNumber(),
-                                        localFile.getAbsolutePath(),
-                                        e.getMessage());
-                                throw e;
-                            } finally {
-                                if (syncService != null) {
-                                    syncService.close();
-                                }
-                            }
-                            return status;
-                        }
-                    };
-            return performDeviceAction(
-                    String.format("pull %s to %s", remoteFilePath, localFile.getAbsolutePath()),
-                    pullAction,
-                    MAX_RETRY_ATTEMPTS);
+            return pullFileInternal(remoteFilePath, localFile);
         } finally {
             long totalTime = System.currentTimeMillis() - startTime;
             InvocationMetricLogger.addInvocationMetrics(
@@ -1338,6 +1303,45 @@ public class NativeDevice implements IManagedTestDevice, IConfigurationReceiver 
         String externalPath = getMountPoint(IDevice.MNT_EXTERNAL_STORAGE);
         String fullPath = new File(externalPath, remoteFilePath).getPath();
         return pullFile(fullPath);
+    }
+
+    protected boolean pullFileInternal(String remoteFilePath, File localFile)
+            throws DeviceNotAvailableException {
+        DeviceAction pullAction =
+                new DeviceAction() {
+                    @Override
+                    public boolean run()
+                            throws TimeoutException, IOException, AdbCommandRejectedException,
+                                    SyncException {
+                        SyncService syncService = null;
+                        boolean status = false;
+                        try {
+                            syncService = getIDevice().getSyncService();
+                            syncService.pullFile(
+                                    interpolatePathVariables(remoteFilePath),
+                                    localFile.getAbsolutePath(),
+                                    SyncService.getNullProgressMonitor());
+                            status = true;
+                        } catch (SyncException e) {
+                            CLog.w(
+                                    "Failed to pull %s from %s to %s. Message %s",
+                                    remoteFilePath,
+                                    getSerialNumber(),
+                                    localFile.getAbsolutePath(),
+                                    e.getMessage());
+                            throw e;
+                        } finally {
+                            if (syncService != null) {
+                                syncService.close();
+                            }
+                        }
+                        return status;
+                    }
+                };
+        return performDeviceAction(
+                String.format("pull %s to %s", remoteFilePath, localFile.getAbsolutePath()),
+                pullAction,
+                MAX_RETRY_ATTEMPTS);
     }
 
     /**
@@ -1451,7 +1455,7 @@ public class NativeDevice implements IManagedTestDevice, IConfigurationReceiver 
     /** {@inheritDoc} */
     @Override
     public boolean doesFileExist(String deviceFilePath) throws DeviceNotAvailableException {
-        return doesFileExist(deviceFilePath, 0);
+        return doesFileExist(deviceFilePath, getCurrentUser());
     }
 
     @Override
@@ -1485,10 +1489,13 @@ public class NativeDevice implements IManagedTestDevice, IConfigurationReceiver 
         long startTime = System.currentTimeMillis();
         try {
             if (isSdcardOrEmulated(deviceFilePath)) {
-                ContentProviderHandler handler = getContentProvider();
-                if (handler != null) {
-                    if (handler.deleteFile(deviceFilePath)) {
-                        return;
+                int currentUser = getCurrentUser();
+                if (currentUser != 0) {
+                    ContentProviderHandler handler = getContentProvider();
+                    if (handler != null) {
+                        if (handler.deleteFile(deviceFilePath)) {
+                            return;
+                        }
                     }
                 }
             }
@@ -1853,65 +1860,72 @@ public class NativeDevice implements IManagedTestDevice, IConfigurationReceiver 
             throws DeviceNotAvailableException {
         long startTime = System.currentTimeMillis();
         try {
+            int currentUser = getCurrentUser();
             if (isSdcardOrEmulated(deviceFilePath)) {
-                ContentProviderHandler handler = getContentProvider();
-                if (handler != null) {
-                    return handler.pullDir(deviceFilePath, localDir);
+                if (currentUser != 0) {
+                    ContentProviderHandler handler = getContentProvider();
+                    if (handler != null) {
+                        return handler.pullDir(deviceFilePath, localDir);
+                    }
                 }
             }
 
-            if (!localDir.isDirectory()) {
-                CLog.e("Local path %s is not a directory", localDir.getAbsolutePath());
-                return false;
-            }
-            if (!doesFileExist(deviceFilePath)) {
-                CLog.e("Device path %s does not exist to be pulled.", deviceFilePath);
-            }
-            if (!isDirectory(deviceFilePath)) {
-                CLog.e("Device path %s is not a directory", deviceFilePath);
-                return false;
-            }
-            FileEntry entryRoot =
-                    new FileEntry(null, deviceFilePath, FileListingService.TYPE_DIRECTORY, false);
-            IFileEntry entry = getFileEntry(entryRoot);
-            Collection<IFileEntry> children = entry.getChildren(false);
-            if (children.isEmpty()) {
-                CLog.i("Device path is empty, nothing to do.");
-                return true;
-            }
-            for (IFileEntry item : children) {
-                if (item.isDirectory()) {
-                    // handle sub dir
-                    File subDir = new File(localDir, item.getName());
-                    if (!subDir.mkdir()) {
-                        CLog.w(
-                                "Failed to create sub directory %s, aborting.",
-                                subDir.getAbsolutePath());
-                        return false;
-                    }
-                    String deviceSubDir = item.getFullPath();
-                    if (!pullDir(deviceSubDir, subDir)) {
-                        CLog.w(
-                                "Failed to pull sub directory %s from device, aborting",
-                                deviceSubDir);
-                        return false;
-                    }
-                } else {
-                    // handle regular file
-                    File localFile = new File(localDir, item.getName());
-                    String fullPath = item.getFullPath();
-                    if (!pullFile(fullPath, localFile)) {
-                        CLog.w("Failed to pull file %s from device, aborting", fullPath);
-                        return false;
-                    }
-                }
-            }
-            return true;
+            return pullDirInternal(deviceFilePath, localDir, currentUser);
         } finally {
             InvocationMetricLogger.addInvocationMetrics(
                     InvocationMetricKey.PULL_DIR_TIME, System.currentTimeMillis() - startTime);
             InvocationMetricLogger.addInvocationMetrics(InvocationMetricKey.PULL_DIR_COUNT, 1);
         }
+    }
+
+    private boolean pullDirInternal(String deviceFilePath, File localDir, int userId)
+            throws DeviceNotAvailableException {
+        if (!localDir.isDirectory()) {
+            CLog.e("Local path %s is not a directory", localDir.getAbsolutePath());
+            return false;
+        }
+        if (!doesFileExist(deviceFilePath, userId)) {
+            CLog.e("Device path %s does not exist to be pulled.", deviceFilePath);
+            return false;
+        }
+        if (!isDirectory(deviceFilePath)) {
+            CLog.e("Device path %s is not a directory", deviceFilePath);
+            return false;
+        }
+        FileEntry entryRoot =
+                new FileEntry(null, deviceFilePath, FileListingService.TYPE_DIRECTORY, false);
+        IFileEntry entry = getFileEntry(entryRoot);
+        Collection<IFileEntry> children = entry.getChildren(false);
+        if (children.isEmpty()) {
+            CLog.i("Device path is empty, nothing to do.");
+            return true;
+        }
+        for (IFileEntry item : children) {
+            if (item.isDirectory()) {
+                // handle sub dir
+                File subDir = new File(localDir, item.getName());
+                if (!subDir.mkdir()) {
+                    CLog.w(
+                            "Failed to create sub directory %s, aborting.",
+                            subDir.getAbsolutePath());
+                    return false;
+                }
+                String deviceSubDir = item.getFullPath();
+                if (!pullDirInternal(deviceSubDir, subDir, userId)) {
+                    CLog.w("Failed to pull sub directory %s from device, aborting", deviceSubDir);
+                    return false;
+                }
+            } else {
+                // handle regular file
+                File localFile = new File(localDir, item.getName());
+                String fullPath = item.getFullPath();
+                if (!pullFileInternal(fullPath, localFile)) {
+                    CLog.w("Failed to pull file %s from device, aborting", fullPath);
+                    return false;
+                }
+            }
+        }
+        return true;
     }
 
     /** Checks whether path is external storage path. */
